✨ 🔧 Restructure: Updated project directory and file structure
