+++ .idea/deploymentTargetSelector.xml
@@ -4,7 +4,7 @@
     <selectionStates>
       <SelectionState runConfigName="app">
         <option name="selectionMode" value="DROPDOWN" />
-        <DropdownSelection timestamp="2024-07-07T06:59:45.835202Z">
+        <DropdownSelection timestamp="2024-07-08T02:21:24.038064Z">
           <Target type="DEFAULT_BOOT">
             <handle>
               <DeviceId pluginId="LocalEmulator" identifier="path=/Users/jinwoo/.android/avd/Galaxy_S7_API_26.avd" />
