+++ app/src/main/java/com/example/week2/data/meal/Meal.kt
@@ -5,18 +5,12 @@ import androidx.room.Entity
 import androidx.room.Index
 import androidx.room.PrimaryKey
 
-enum class MealTime {
-    BREAKFAST,
-    LUNCH,
-    DINNER,
-    SNACK,
-    NIGHT_SNACK
-}
+
 
 @Entity(tableName = "meal_table", indices = [Index(value = ["date"])])
 data class Meal(
     @PrimaryKey(autoGenerate = true) val id: Int = 0,
-    @ColumnInfo(name = "meal_time") val mealTime: MealTime,
+    @ColumnInfo(name = "meal_time") val mealTime: String,
     @ColumnInfo(name = "meal_name") val mealName: String,
     @ColumnInfo(name = "price") val price: Int,
     @ColumnInfo(name = "date") val date: String,
