+++ app/src/main/java/com/example/week2/data/word/WordRepository.kt
@@ -1,4 +1,4 @@
-package com.example.week2
+package com.example.week2.data.word
 
 /*
  * Copyright (C) 2017 Google Inc.
