+++ app/src/main/java/com/example/week2/data/word/WordsApplication.kt
@@ -1,6 +1,8 @@
-package com.example.week2
+package com.example.week2.data.word
 
 import android.app.Application
+import com.example.week2.data.AppRoomDatabase
+import com.example.week2.MealRepository
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.SupervisorJob
