+++ app/src/main/java/com/example/week2/ui/activity/FriendActivity.kt
