+++ app/src/main/java/com/example/week2/ui/activity/MealActivity.kt
@@ -11,6 +11,8 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.widget.Toolbar
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
+import com.example.week2.data.word.WordsApplication
+import com.example.week2.ui.adapter.MealListAdapter
 import com.google.android.material.floatingactionbutton.FloatingActionButton
 
 class MealActivity : AppCompatActivity() {
