+++ app/src/main/res/values/strings.xml
@@ -17,4 +17,11 @@
     <string name="word_empty_not_saved">Word not saved because it is empty.</string>
     <string name="meal_empty_not_saved">식비 지출이 기록되지 않았습니다.</string>
     <string name="add_word">Add word</string>
+    <string-array name="meal_times_array">
+        <item>아침</item>
+        <item>점심</item>
+        <item>저녁</item>
+        <item>야식</item>
+        <item>간식</item>
+    </string-array>
 </resources>
